feat: connect db
