+++ src/app/api/product/route.js
@@ -1,0 +1,7 @@
+import { connectDB } from '@/connectDB';
+import { NextResponse } from 'next/server';
+
+export async function GET() {
+  connectDB();
+  return NextResponse.json({ products: { name: 'Product 1', price: 100 } });
+}
