+++ src/connectDB.js
@@ -1,0 +1,13 @@
+import mongoose from 'mongoose';
+
+export const connectDB = async () => {
+  try {
+    const conn = await mongoose.connect(process.env.MONGO_URI, {
+      useUnifiedTopology: true,
+      useNewUrlParser: true,
+    });
+    console.log(`MongoDB Connected: ${conn.connection.db.databaseName}`);
+  } catch (error) {
+    console.error(`Error: ${error.message}`);
+  }
+};
